feat(proj2): add func that defines a list of projects
