+++ .gitignore
@@ -1,3 +1,4 @@
 .venv
 .pytest_cache
 *.pytest_cache
+__pycache__
